+++ History.md
@@ -16,7 +16,7 @@
   * Log remote IP
   * Allow options.stackTraceLimit to be falsy
   * Remove client.version property
-  * Control log level of client with options.level (replaces options.silent)
+  * Control log level of client with options.clientLogLevel (replaces options.silent)
   * Allow handleUncaughtExceptions to be called multiple times
   * Allow the severity level of exceptions to be set in options
   * Allow all options to be set via environment variables